Device names should handle unicode gracefully now
portaudio returns names in whatever encoding the system sees fit. On
most systems, for most host APIs, this is unicode. However, for
DirectSound on Windows, it is the current code page. This is taken
into account when decoding device names now.

Also, other decoding errors are ignored instead of crashing
pysoundcard.